Reduce duplicate code with custom hooks Fixes #23
